tools/requirements.txt
pdfminer.six
duckduckgo-search
newspaper3k